Board removes bullets from array, that have left the screen.
removeOffCanvasBullets function was added, which iterates through the collection of bullets
within the board object to determine which are off of the screen. This could be chnage to only look
at the last bullet in the array and remove it, once it has reached the end. However, I decided to
iterate incase we decide to add functionality where multiple bullets can be fired at the same time.

Next step will be to try and change the rate of bullet fire.