Added movement to meteors horizontally.
Changed the velocity on the x position for the meteor to be between -2 and 2, and updated changeImpactMeteorAndObjectStatuses function
with updated logic for removing meteor when it exits the screen horizontally. This was done so that we do not have to iterate
through the meteors a second time and potentially slow the program down to remove meteors. The code logic will need to be refactored
as it violates the SRP principle. But for now it does the thing.

Next step besides refactoring will be to add the delay on bullet firing and refactoring.